refactor: move more map related components to the components package and some TS conversion
