+++ src/components/map/interactions/ShowContextMenu.js
@@ -22,9 +22,8 @@ import React from 'react';
 import { OLPropTypes, withMap } from '@collmot/ol-react';
 import { createOLInteractionComponent } from '@collmot/ol-react/lib/interaction';
 
+import * as Condition from '~/components/map/conditions';
 import { euclideanDistance2D } from '~/utils/math';
-
-import * as Condition from '../conditions';
 
 /**
  * OpenLayers interaction (for right click by default) that overwrites the
