+++ src/components/map/interactions/TrackNearestFeature.js
