+++ src/features/hotkeys/actions.js
@@ -1,14 +1,14 @@
 import isNil from 'lodash-es/isNil';
 
-import { dismissAlerts } from '~/features/alert/slice';
+import { copyDisplayedCoordinatesToClipboard } from '~/components/map/utils';
 import { hasPendingAudibleAlerts } from '~/features/alert/selectors';
+import { dismissAlerts } from '~/features/alert/slice';
 import { clearSelection } from '~/features/map/selection';
 import { getMissionMapping } from '~/features/mission/selectors';
 import { showNotification } from '~/features/snackbar/actions';
 import { setSelectedUAVIds } from '~/features/uavs/actions';
 import { getUAVById } from '~/features/uavs/selectors';
 import { scrollUAVListItemIntoView } from '~/utils/navigation';
-import { copyDisplayedCoordinatesToClipboard } from '~/views/map/utils';
 
 import { getPendingUAVId, isPendingUAVIdOverlayVisible } from './selectors';
 import { setPendingUAVId, startPendingUAVIdTimeout } from './slice';
